% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{source_code}
\alias{source_code}
\title{Source a block of text containing C code}
\usage{
source_code(x, includes = NULL, no_remap = TRUE, show = FALSE)
}
\arguments{
\item{x}{\verb{[character(1)]}

A block of C code to compile.}

\item{includes}{\verb{[NULL / character()]}

Extra includes to add manually. By default, \code{R.h} and \code{Rinternals.h} are
included. Specify more includes with their file name. For example, to
include \verb{#include <Rdefines.h>} you just need to specify \code{"Rdefines.h"}.}

\item{no_remap}{\verb{[logical(1)]}

Should \verb{#define R_NO_REMAP} be defined?}

\item{show}{\verb{[logical(1)]}

Should the output of compiling the source code with \verb{R CMD SHLIB} be shown?}
}
\value{
A named list containing the functions specified for export.
}
\description{
\code{source_code()} will parse through \code{x} looking for functions tagged with
\verb{// [[ export() ]]} and will compile the code block and export those
functions to the R side.
}
\examples{
code <- "
  static SEXP helper(SEXP x) {
    return x;
  }

  // [[ export() ]]
  SEXP fn1(SEXP x) {
    return helper(x);
  }

  // [[ export() ]]
  SEXP fn2(SEXP x, SEXP y) {
    double result = REAL(x)[0] + REAL(y)[0];
    return Rf_ScalarReal(result);
  }
"

sourced <- source_code(code)

sourced$fn1(1)
sourced$fn2(1, 2)
}
